+++ src/Service/YamlFilesPathService.php
@@ -13,10 +13,14 @@ class YamlFilesPathService
     /**
      * @param string[] $pathToDirsOrFiles
      * @param string[] $excludedPaths
+     * @param bool $includeSkippedFilesToPath
      * @return string[]
      */
-    public static function getPathToYamlFiles(array $pathToDirsOrFiles, array $excludedPaths)
-    {
+    public static function getPathToYamlFiles(
+        array $pathToDirsOrFiles,
+        array $excludedPaths,
+        $includeSkippedFilesToPath = false
+    ) {
         $pathToFiles = [];
         foreach ($pathToDirsOrFiles as $pathToDirOrFile) {
             if (is_file($pathToDirOrFile)) {
@@ -24,9 +28,7 @@ class YamlFilesPathService
                 continue;
             }
 
-            $recursiveDirectoryIterator = new RecursiveDirectoryIterator($pathToDirOrFile);
-            $recursiveDirectoryFilterIterator = new RecursiveDirectoryFilterIterator($recursiveDirectoryIterator, $excludedPaths);
-            $recursiveIteratorIterator = new RecursiveIteratorIterator($recursiveDirectoryFilterIterator);
+            $recursiveIteratorIterator = self::getRecursiveIteratorIterator($pathToDirOrFile, $excludedPaths, $includeSkippedFilesToPath);
             $regexIterator = new RegexIterator($recursiveIteratorIterator, '/^.+\.(ya?ml(\.dist)?)$/i', RecursiveRegexIterator::GET_MATCH);
 
             foreach ($regexIterator as $pathToFile) {
@@ -37,5 +39,28 @@ class YamlFilesPathService
         $pathToFiles = str_replace('\\', '/', $pathToFiles);
 
         return array_unique($pathToFiles);
+    }
+
+    /**
+     * @param string $pathToDir
+     * @param string[] $excludedPaths
+     * @param bool $includeSkippedFilesToPath
+     * @return \RecursiveIteratorIterator
+     */
+    private static function getRecursiveIteratorIterator(
+        $pathToDir,
+        array $excludedPaths,
+        $includeSkippedFilesToPath = false
+    ) {
+        $recursiveDirectoryIterator = new RecursiveDirectoryIterator($pathToDir);
+
+        if ($includeSkippedFilesToPath) {
+            $recursiveIteratorIterator = new RecursiveIteratorIterator($recursiveDirectoryIterator);
+        } else {
+            $recursiveDirectoryFilterIterator = new RecursiveDirectoryFilterIterator($recursiveDirectoryIterator, $excludedPaths);
+            $recursiveIteratorIterator = new RecursiveIteratorIterator($recursiveDirectoryFilterIterator);
+        }
+
+        return $recursiveIteratorIterator;
     }
 }